Created LED show sequence
Sequencing done in TouchDesigner to keep the majority of logic centralized. The TD project has been changed so that the audio track for the show is mapped to the timeline. Booleans are triggered as the frame count increases, sending messages over osc to Chataigne to control what light mode is active as the song progresses. Just need to press "Space" to start playing the sequence in TD. Also make sure the "Audio Device Out" CHOP is disabled during the show